+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * Choose whether to sort sensisitivities while plotting.
 * Added Sim.plot() method
 * Export invar nums to a csv or json file with sim.exportInVarNums
+* Import outvals from a csv or json with sim.importOutVals, and convert to outvars
 ### Changed    
 * Prevent overwriting existing InVars or OutVals with an already used name.
 ### Removed